+++ npm-packages/shared-utils/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @pulse-editor/shared-utils
 
+## 0.1.1-beta.76
+
+### Patch Changes
+
+- Fix type error
+
+## 0.1.1-beta.75
+
+### Patch Changes
+
+- Refactor app skill actions
+
 ## 0.1.1-beta.74
 
 ### Patch Changes